chore: refactor the casbin to instantiate new enforcer and addapter separately
